+++ docs/change_log.rst
@@ -2,7 +2,7 @@
 ===========================
 2018.10.27 v1.1.2
 ---------------------------
-* `#80 <https://github.com/chinapnr/fishbase/issues/80>`_, common, add function :meth:`fish_common.find_common_between_dicts`, doc and unittest;
+* `#80 <https://github.com/chinapnr/fishbase/issues/80>`_, common, add function :meth:`fish_common.find_same_between_dicts`, doc and unittest;
 * `#99 <https://github.com/chinapnr/fishbase/issues/99>`_, common, add function :meth:`fish_common.GetMD5.hmac_md5`, doc and unittest;